+++ node/clients/index.ts
@@ -2,6 +2,8 @@ import { IOClients } from '@vtex/api'
 
 import { Search } from './search'
 import { Catalog } from './catalog'
+import { CustomPricing } from './customPricing'
+import { CustomStockAvailability } from './customStockAvailability'
 
 export class Clients extends IOClients {
   public get search(): any {
@@ -10,5 +12,13 @@ export class Clients extends IOClients {
 
   public get catalog() {
     return this.getOrSet('catalog', Catalog)
+  }
+
+  public get customPricing() {
+    return this.getOrSet('customPricing', CustomPricing)
+  }
+
+  public get customStockAvailability() {
+    return this.getOrSet('customStockAvailability', CustomStockAvailability)
   }
 }