+++ react/components/ReviewBlock.tsx
@@ -16,6 +16,8 @@ import { useCssHandles } from 'vtex.css-handles'
 
 import { GetText, validateQuantity } from '../utils'
 import { keyValuePairsToString } from '../utils/performanceDataProcessing'
+import ItemPricing from './ItemPricing'
+import StockAvailability from './StockAvailability'
 // eslint-disable-next-line @typescript-eslint/ban-ts-comment
 // @ts-ignore
 // import getRefIdTranslation from '../queries/refids.gql'
@@ -194,6 +196,11 @@ const CSS_HANDLES = [
   'inStockMessage',
   'outOfStockMessage',
   'stockAvailabilityMessage',
+  'stockAvailabilityValue',
+  'productPageLink',
+  'orderedQuantity',
+  'orderedQuantityLabel',
+  'orderedQuantityValue',
 ]
 
 const ReviewBlock: FunctionComponent<WrappedComponentProps & any> = ({
@@ -408,9 +415,7 @@ const ReviewBlock: FunctionComponent<WrappedComponentProps & any> = ({
           availableQuantity: getAvailableQuantity(item),
           price: getPrice(item),
           vtexSku: vtexSku(item),
-          totalQuantity:
-            (item.sku ? mappedRefId[item.sku]?.unitMultiplier : '1') *
-            item.quantity,
+          totalQuantity: item.quantity,
           error: errorMsg(item),
           availability: getAvailability(item),
           productName: itm?.productName,
@@ -622,20 +627,41 @@ const ReviewBlock: FunctionComponent<WrappedComponentProps & any> = ({
                     </span>
                   </div>
                 </div>
-                <div className={`${styles.productLink} flex justify-end w-100`}>
-                  <a
-                    className={`${styles.productDetailsLink} flex-column`}
-                    href={`${rowData?.linkText}/p`}
-                    target="_blank"
-                    rel="noreferrer"
-                  >
-                    {intl.formatMessage(messages.goToProductPage)}
-                  </a>
+                <div
+                  className={`${styles.productLink} flex justify-between w-100`}
+                >
+                  <div className={`${styles.orderedQuantity} truncate`}>
+                    <span className={`${styles.orderedQuantityLabel} mr3`}>
+                      Ordered Quantity:
+                    </span>
+                    <span className={`${styles.orderedQuantityValue}`}>
+                      {rowData.quantity}
+                    </span>
+                  </div>
+                  <div className={`${styles.productPageLink} ml3`}>
+                    <a
+                      className={`${styles.productDetailsLink} flex-column`}
+                      href={`${rowData?.linkText}/p`}
+                      target="_blank"
+                      rel="noreferrer"
+                    >
+                      {intl.formatMessage(messages.goToProductPage)}
+                    </a>
+                  </div>
                 </div>
               </div>
 
               <div className={`${styles.tableCol2} flex w-40 pa3`}>
-                <div className={`${styles.tableCol2Col1} w-40`} />
+                <div className={`${styles.tableCol2Col1} w-40`}>
+                  {targetSystem === 'JDE' ? (
+                    <ItemPricing
+                      itemNumber={rowData?.sku}
+                      customerNumber={customerNumber}
+                    />
+                  ) : (
+                    <div />
+                  )}
+                </div>
                 <div className={`${styles.tableCol2Col2} w-60`}>
                   <div
                     className={`${styles.itemUom} flex flex-row justify-between`}
@@ -681,7 +707,7 @@ const ReviewBlock: FunctionComponent<WrappedComponentProps & any> = ({
               </div>
               <div className={`${styles.tableCol3} flex flex-column w-20 pa3`}>
                 <div
-                  className={`${styles.stockAvailabilityMessage} flex justify-center w-100`}
+                  className={`${styles.stockAvailabilityMessage} flex w-100`}
                 >
                   <Tooltip label={statusMessage}>
                     {rowData?.availability === 'available' ||
@@ -696,6 +722,16 @@ const ReviewBlock: FunctionComponent<WrappedComponentProps & any> = ({
                       </span>
                     )}
                   </Tooltip>
+                </div>
+                <div className={`${styles.stockAvailabilityValue} w-40`}>
+                  {targetSystem === 'JDE' ? (
+                    <StockAvailability
+                      itemNumber={rowData?.sku}
+                      customerNumber={customerNumber}
+                    />
+                  ) : (
+                    <div />
+                  )}
                 </div>
               </div>
               <div className={`${styles.deleteRowBtn} absolute right-0 top-0`}>