+++ react/utils/index.ts
@@ -112,3 +112,11 @@ export const validateQuantity = (minQty: number, unit: number, qty: number) => {
     ? minQty + (unit - (minQty % unit))
     : actualQty
 }
+
+export const getFormattedDate = (date: Date) => {
+  let year = date.getFullYear();
+  let month = (1 + date.getMonth()).toString().padStart(2, '0');
+  let day = date.getDate().toString().padStart(2, '0');
+
+  return month + '/' + day + '/' + year;
+}